requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.9.0
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
praw==7.8.1
prawcore==2.4.0
protobuf==6.33.0
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
tenacity==9.1.2
textblob==0.19.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
websocket-client==1.9.0
